+++ src/api/controllers/comment.controller.ts
@@ -1,4 +1,4 @@
-import CommentService from "../services/comment/CommentService.js";
+import CommentService from "../services/comment/comment.service.js";
 import type { Request, Response, NextFunction } from "express";
 import type { CommentDto } from "../types/dtos/comment.dto.js";
