+++ src/api/controllers/image.controller.ts
