+++ src/api/controllers/product.controller.ts
@@ -1,27 +1,27 @@
-import ArticleService from "../services/article/ArticleService.js";
+import ProductService from "../services/product/product.service.js";
 import jwt from "jsonwebtoken";
-import type { Request, Response, NextFunction } from "express";
 import { ACCESS_TOKEN_SECRET } from "../libs/constants.js";
+import type { Request, Response, NextFunction } from "express";
 import type { CustomError } from "src/api/types/error.js";
-import type { ArticleDto } from "../types/dtos/article.dto.js";
+import type { ProductDto } from "../types/dtos/product.dto.js";
 
-const ArticleController = {
-  async createArticle(req: Request, res: Response, next: NextFunction) {
+const ProductController = {
+  async createProduct(req: Request, res: Response, next: NextFunction) {
     try {
       const { id: userId } = req.user;
-      const articleData: ArticleDto = req.body;
-      const newArticle = await ArticleService.createArticle(articleData, userId);
+      const productData: ProductDto = req.body;
+      const newProduct = await ProductService.createProduct(productData, userId);
 
-      res.status(201).json(newArticle);
+      res.status(201).json(newProduct);
     } catch (err) {
       next(err);
     }
   },
 
-  async findUniqueArticle(req: Request, res: Response, next: NextFunction) {
+  async findUniqueProduct(req: Request, res: Response, next: NextFunction) {
     try {
       const { id } = req.params;
-      const articleId = Number(id);
+      const productId = Number(id);
 
       let userId = null;
       const token = req.cookies.accessToken;
@@ -39,56 +39,62 @@ const ArticleController = {
           console.error("토큰 검증 오류:", err);
         }
       }
-      const article = await ArticleService.findUniqueArticle(articleId, userId);
-      res.status(200).json(article);
+
+      const product = await ProductService.findUniqueProduct(productId, userId);
+      res.status(200).json(product);
     } catch (err) {
       next(err);
     }
   },
 
-  async updateArticle(req: Request, res: Response, next: NextFunction) {
+  async patchProduct(req: Request, res: Response, next: NextFunction) {
     try {
-      const { id: userId } = req.user;
       const { id } = req.params;
-      const updateData: ArticleDto = req.body;
-      const article = await ArticleService.updateArticle(Number(id), updateData, userId);
-      res.status(200).json(article);
+      const { id: userId } = req.user;
+      const updateData: ProductDto = req.body;
+
+      const product = await ProductService.patchProduct(Number(id), updateData, userId);
+      if (!product) {
+        return res.status(404).json({ error: "수정할 상품이 없음" });
+      }
+      res.status(200).json(product);
     } catch (err) {
       next(err);
     }
   },
 
-  async deleteArticle(req: Request, res: Response, next: NextFunction) {
+  async deleteProduct(req: Request, res: Response, next: NextFunction) {
     try {
-      const { id: userId } = req.user;
       const { id } = req.params;
-      await ArticleService.deleteArticle(Number(id), userId);
+      const { id: userId } = req.user;
+      await ProductService.deleteProduct(Number(id), userId);
+
       res.status(204).json({ success: "상품 삭제 성공" });
     } catch (err) {
       next(err);
     }
   },
 
-  async findManyArticle(req: Request, res: Response, next: NextFunction) {
+  async findManyProduct(req: Request, res: Response, next: NextFunction) {
     try {
-      const { offset = 0, limit = 10, order, keyword } = req.query;
+      const { offset = 0, limit = 10, order = "recent", keyword } = req.query;
 
       const finalOffset = Number(offset) || 0;
       const finalLimit = Number(limit) || 10;
       const finalOrder = typeof order === "string" ? order : "recent";
       const finalKeyword = typeof keyword === "string" ? keyword : undefined;
 
-      const articles = await ArticleService.findManyArticle({
+      const products = await ProductService.findManyProduct({
         offset: finalOffset,
         limit: finalLimit,
         order: finalOrder,
         ...(finalKeyword && { keyword: finalKeyword }),
       });
-      res.status(200).json(articles);
+      res.status(200).json(products);
     } catch (err) {
       next(err);
     }
   },
 };
 
-export default ArticleController;
+export default ProductController;
