+++ src/api/routes/mypage.router.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import authenticate from "../middlewares/authenticate.js";
-import MypageController from "../controllers/MypageController.js";
+import MypageController from "../controllers/mypage.controller.js";
 import validate from "../middlewares/validate.js";
 import { updateUserSchema, updatePasswordSchema } from "../types/dtos/mypage.dto.js";
