+++ src/api/services/auth/auth.service.ts
@@ -3,7 +3,7 @@ import { hashing, compareWords } from "../../libs/hashing.js";
 import { generateTokens } from "../../libs/token.js";
 import { REFRESH_TOKEN_SECRET } from "../../libs/constants.js";
 import type { CustomError } from "src/api/types/error.js";
-import * as AuthRepository from "../../repositories/AuthRepository.js";
+import * as AuthRepository from "../../repositories/auth.repository.js";
 import type { SignupDto, LoginDto } from "../../types/dtos/auth.dto.js";
 
 const AuthService = {
