+++ src/api/services/like/like.service.ts
@@ -1,5 +1,5 @@
 import type { LikeData } from "../../types/like.js";
-import * as LikeRepository from "../../repositories/LikeRepository.js";
+import * as LikeRepository from "../../repositories/like.repository.js";
 import type { Prisma } from "@prisma/client";
 
 const LikeService = {
