+++ src/api/services/product/product.service.ts
@@ -1,6 +1,6 @@
 import type { CustomError } from "../../types/error.js";
 import type { CreateProductData, UpdateProductData, FindManyProductParams } from "../../types/product.js";
-import * as ProductRepository from "../../repositories/ProductRepository.js";
+import * as ProductRepository from "../../repositories/product.repository.js";
 import type { ProductDto } from "../../types/dtos/product.dto.js";
 
 const ProductService = {
