+++ src/external/classes/article.ts
