+++ src/external/classes/product.ts
